tests: Rename `test_stream.js` to `example_stream.js`.
