+++ web/tests/example2.test.js
@@ -2,9 +2,9 @@
 
 const {strict: assert} = require("assert");
 
+const {make_stream} = require("./lib/example_stream");
 const {zrequire} = require("./lib/namespace");
 const {run_test} = require("./lib/test");
-const {make_stream} = require("./lib/test_stream");
 
 // Hopefully the basic patterns for testing data-oriented modules
 // are starting to become apparent.  To reinforce that, we will present
