+++ web/tests/example3.test.js
@@ -2,9 +2,9 @@
 
 const {strict: assert} = require("assert");
 
+const {make_stream} = require("./lib/example_stream");
 const {zrequire} = require("./lib/namespace");
 const {run_test} = require("./lib/test");
-const {make_stream} = require("./lib/test_stream");
 const {realm} = require("./lib/zpage_params");
 
 // In the Zulip app you can narrow your message stream by topic, by
