+++ web/tests/lib/example_stream.js
